Introduce hubs-systems system
