+++ src/hub.html
@@ -30,7 +30,7 @@
 <body>
     <a-scene
         loading-screen="enabled: false"
-        nextframe
+        hubs-systems
         class="grab-cursor"
         renderer="antialias: true; colorManagement: true; sortObjects: true; physicallyCorrectLights: true; alpha: false; webgl2: true; multiview: false;"
         shadow="type: pcfsoft"
